feat. regional prices endpoint
Created a regional price endpoint that takes the ip and returns the country, the default prices and the regional prices if they exist.
Corrected the price_weights csv file so that the prices can be parsed to a number. Changed the default prices test so that it's going to expect numbers.
Deleted unnescessary endpoints.